internal refactoring
