+++ pyLinguist/tools.py
